Warning cleanup
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23673 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/plots/ParallelCoordinates/ParallelCoordinatesAttributes.C
@@ -1358,7 +1358,7 @@ void
 ParallelCoordinatesAttributes::DeleteAxis(const std::string &axisName_,
                                           int minAxisCount)
 {
-    if (scalarAxisNames.size() <= (size_t)minAxisCount) return;
+    if ((int)scalarAxisNames.size() <= minAxisCount) return;
 
     std::string newAxisName = axisName_;
